fix hero rendering on linux

+++ src/utils/animations.cpp
@@ -22,8 +22,8 @@ namespace df {
         while (elapsedTime >= frameDuration) {
             elapsedTime -= frameDuration;
             currentFrameIndex++;
-            if (currentFrameIndex >= static_cast<int>(frames.size())) {
-                currentFrameIndex = loop ? 0 : static_cast<int>(frames.size()) - 1;
+            if (currentFrameIndex >= frames.size()) {
+                currentFrameIndex = loop ? 0 : frames.size() - 1;
             }
         }
     }
@@ -45,7 +45,6 @@ namespace df {
     }
 
 
- 
 
     void Animation::setLoop(bool l) { loop = l; }
     bool Animation::isLooping() const { return loop; }
